requirements.txt
annotated-types = 0.7.0  
anyio = 4.9.0  
bcrypt = 4.3.0  
beautifulsoup4 = 4.13.4  
certifi = 2025.4.26  
charset-normalizer = 3.4.2  
click = 8.1.8  
colorama = 0.4.6  
dnspython = 2.7.0  
email_validator = 2.2.0  
fastapi = 0.115.12  
greenlet = 3.2.2  
h11 = 0.16.0  
idna = 3.10  
passlib = 1.7.4  
pip = 25.1.1  
pydantic = 2.11.4  
pydantic_core = 2.33.2  
PyJWT = 2.10.1  
python-dotenv = 1.1.0  
python-multipart = 0.0.20  
requests = 2.32.3  
sniffio = 1.3.1  
soupsieve = 2.7  
SQLAlchemy = 2.0.40  
starlette = 0.46.2  
typing_extensions = 4.13.2  
typing-inspection = 0.4.0  
urllib3 = 2.4.0  
uvicorn = 0.34.2  
